Go through session repository for VI Session access and initialization. (#49)
* Use Session defined in server_utilities.proto.

* Use access_session for Sessions on requests.

* Introduce CleanupVISession protected method.

* Special implementation for init session methods.

* Remove unnecessary include.

* Set status on response for init session methods.

* Rely on config's 'init_function' instead of marking function in functions.py.

* Use make_tuple like in tests.

* Call close through library wrapper in lambda directly.

* Rename lambda variables.

* More minor cleanup.

* Just set id for Session response from init.

* Typedef init func as well for session_repository.

* Close session goes through session_repository's remove_session.

* Clean up how to get session out param name.

* Remove extra parentheses around request_snippet.

* Checkpoint 1 refactor (rebase and squash this one)

* Check which field was set on Session.

* Move defs to bottom of file.

* Separate request input parameters into def.

* Separate out init and simple method body's out into defs.

* Initialize session_id to 0.

* Separate out initializing enums with mapping into separate def.

* Factor out setting response values to def.

* Rename top level def's to have better names.

* Rely on default scalar values for id and name on session.

* Indent body of multi-line lambda.

* Make session_name gotten from the request a const std::string&.

* Add blank line after closing namespace brackets.

* Fix up end of file to end in blank line.

* Refactor to sanitize params only once and pass more info to top level defs.

* Make session's id field a uint32 instead of fixed32.

+++ source/tests/unit/hardware/grpc/core_service_tests.cpp
@@ -361,7 +361,7 @@ TEST(CoreServiceTests, ReservationAndSession_ResetServer_UnreservesAndRemovesSes
   ni::hardware::grpc::internal::DeviceEnumerator device_enumerator;
   ni::hardware::grpc::CoreService service(&session_repository, &device_enumerator);
   std::string session_name = "session_name";
-  uint64_t named_session_id;
+  uint32_t named_session_id;
   int status = session_repository.add_session(
       session_name,
       []() { return std::make_tuple(0, 42); },
